requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask==3.0.0
sqlalchemy==2.0.23

